dynmod_loader/requirement.txt
uvicorn==0.29.0
fastapi==0.111.0
yadisk==2.1.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0